fix: accessibility errors
