+++ src/components/_plans/Calendar/Calendar.tsx
@@ -45,7 +45,6 @@ const Calendar = () => {
         isMonthsView={isMonthsView}
         isYearsView={isYearsView}
       />
-      <img src="" />
       <div onWheel={handleMouseWheel}>{getRenderList()}</div>
     </div>
   );
