+++ src/components/_plans/Calendar/components/CalendarHeader/CalendarHeader.tsx
@@ -3,9 +3,6 @@ import { FC } from 'react';
 import Arrows from './components/Arrows/Arrows';
 import CurrentView from './components/CurrentView/CurrentView';
 
-import { useAppDispatch } from 'store/hooks';
-import { calendarActions } from 'store/slices/calendar.slice';
-
 import mergeClasses from 'utils/mergeClasses';
 
 import classes from './CalendarHeader.module.css';
@@ -18,23 +15,9 @@ type Props = {
 };
 
 const CalendarHeader: FC<Props> = ({ month, year, isMonthsView, isYearsView }) => {
-  const dispatch = useAppDispatch();
-
-  const handleViewClick = () => {
-    if (isMonthsView) {
-      dispatch(calendarActions.setIsMonthsView(false));
-      dispatch(calendarActions.setIsYearsView(true));
-    } else if (!isYearsView) {
-      dispatch(calendarActions.setIsMonthsView(true));
-    }
-  };
-
   return (
     <div className={classes.root}>
-      <div
-        className={mergeClasses(classes.dateInformation, classes.headerElement)}
-        onClick={handleViewClick}
-      >
+      <div className={mergeClasses(classes.dateInformation, classes.headerElement)}>
         <CurrentView
           isMonthsView={isMonthsView}
           isYearsView={isYearsView}
