+++ src/components/_plans/Calendar/components/CalendarHeader/components/CurrentView/CurrentView.tsx
@@ -1,6 +1,7 @@
 import { FC } from 'react';
 
-import { useAppSelector } from 'store/hooks';
+import { useAppDispatch, useAppSelector } from 'store/hooks';
+import { calendarActions } from 'store/slices/calendar.slice';
 
 import betterAt from 'utils/betterAt';
 
@@ -15,22 +16,37 @@ type Props = {
 
 const CurrentView: FC<Props> = ({ isMonthsView, isYearsView, month, year }) => {
   const { highlightedYears } = useAppSelector(state => state.calendar);
+  const dispatch = useAppDispatch();
 
   // TODO: think about moving it to some custom format function so it will be easier later to add multi language support
   const visibleMonth = new Intl.DateTimeFormat('en-US', { month: 'long' }).format(
     new Date(0, month),
   );
+
+  const handleViewClick = () => {
+    if (isMonthsView) {
+      dispatch(calendarActions.setIsMonthsView(false));
+      dispatch(calendarActions.setIsYearsView(true));
+    } else if (!isYearsView) {
+      dispatch(calendarActions.setIsMonthsView(true));
+    }
+  };
+
   const renderViewContent = () => {
     if (!isMonthsView && !isYearsView) {
       return (
-        <button className={classes.calendarView} aria-label={`${visibleMonth} ${year}`}>
+        <button
+          className={classes.calendarView}
+          aria-label={`${visibleMonth} ${year}`}
+          onClick={handleViewClick}
+        >
           <span className={classes.month}>{visibleMonth}</span>
           <span>{year}</span>
         </button>
       );
     } else if (isMonthsView) {
       return (
-        <button className={classes.calendarView} aria-label={`${year}`}>
+        <button className={classes.calendarView} aria-label={`${year}`} onClick={handleViewClick}>
           <span>{year}</span>
         </button>
       );
