+++ src/components/_plans/EventsList/components/Event/Event.tsx
@@ -1,4 +1,4 @@
-import { FC, useState } from 'react';
+import { FC } from 'react';
 
 import { EventData } from 'types/store/plansState.type';
 
@@ -7,20 +7,8 @@ import classes from './Event.module.css';
 type Props = Omit<EventData, 'id'>;
 
 const Event: FC<Props> = ({ title, timeFrom, timeTo, location }) => {
-  const [isPressed, setIsPressed] = useState(false);
-
-  const handleEventClick = () => {
-    setIsPressed(prev => !prev);
-  };
-
   return (
-    <div
-      className={classes.root}
-      role="button"
-      aria-pressed={isPressed}
-      tabIndex={0}
-      onClick={handleEventClick}
-    >
+    <div className={classes.root}>
       <div className={classes.container}>
         <div className={classes.time}>
           <p className={classes.timeFrom}>{timeFrom}</p>
